[Refactor] : login, register module merge to auth module
Git : #44
