+++ eb_fast_api/service/notification/testings/mock_notification_provider.py
@@ -2,10 +2,14 @@ from unittest.mock import patch
 from eb_fast_api.service.notification.sources.notification_provider import (
     NotificationScheduleProvider,
 )
+from datetime import datetime
 
 
 def patcher_get_schedule():
-    def mock_get_schedule(self):
+    def mock_get_schedule(
+        self,
+        now: datetime,
+    ):
         cur_noti_schedule = self.pop()
         return [cur_noti_schedule]
